pyproject.toml
[tool.poetry]
name = "leapyear-api"
version = "0.1.0"
description = "Django App using api to verify leap year"
authors = ["ProxTricky"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.3"
Django = "4.2"
djangorestframework = "^3.15.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
